Commit on 20180312 branch master 879129508

+++ scripts/files/working/necessaryfunctions.sh
@@ -209,9 +209,12 @@ makefinishsetup ()
 	printf "\n\033[1;32m==> \033[0m"
 	locale-gen ||:
 	printf "\n\033[1;32m==> \033[1;37mPatching \033[1;32m/bin/makepkg\033[1;37m...\n\n\033[0m"
-	patch -n -i /root/bin/makepkg.diff -o makepkg /bin/makepkg ||: 
+	if [ -e \$HOME/makepkg.diff ];then
+	patch -n -i \$HOME/makepkg.diff -o makepkg /bin/makepkg ||: 
 	mv makepkg /bin/makepkg ||: 
-	rm /root/bin/makepkg.diff ||: 
+	else
+	printf "\nPatch \$HOME/makepkg.diff not found.\n"
+	fi
 	printf "\n\033[1;32m==> \033[1;37mPatching \033[1;32m/bin/makepkg\033[1;37m: \033[1;32mDONE\n\n\033[1;32m==> \033[1;37mRunning \033[1;32mtzselect\033[1;37m...\n\n\033[1;34mAdd the \033[1;32mtzselect\033[1;34m output code to \033[1;32m.bash_profile\033[1;34m so the system time in Arch Linux for future sessions will be set correctly.\n\n\033[0m"
 	tzselect
 	printf "\n"
@@ -428,7 +431,6 @@ touchupsys ()
 	addgcm
 	addgp
 	addgpl
-	addmakepkgdiff
 	addmotd
 	addprofile 
 	addresolvconf
